--- examples/simple_notepad/frontend/css/style.css
@@ -1,207 +1,0 @@
-/* Simple Notepad Application Styles */
-:root {
-    --primary-color: #4a6fa5;
-    --secondary-color: #6b8cae;
-    --accent-color: #166088;
-    --dark-color: #333;
-    --light-color: #f8f9fa;
-    --danger-color: #dc3545;
-    --success-color: #28a745;
-    --border-color: #dee2e6;
-    --shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-}
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-body {
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    line-height: 1.6;
-    color: var(--dark-color);
-    background-color: var(--light-color);
-    height: 100vh;
-    overflow: hidden;
-}
-
-.container {
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-}
-
-header {
-    background-color: var(--primary-color);
-    color: white;
-    padding: 1rem;
-    text-align: center;
-}
-
-header h1 {
-    margin-bottom: 0.2rem;
-}
-
-header p {
-    opacity: 0.8;
-    font-size: 0.9rem;
-}
-
-main {
-    flex: 1;
-    overflow: hidden;
-    padding: 1rem;
-}
-
-.notes-container {
-    display: flex;
-    height: 100%;
-    border: 1px solid var(--border-color);
-    border-radius: 5px;
-    overflow: hidden;
-    box-shadow: var(--shadow);
-}
-
-.notes-sidebar {
-    width: 250px;
-    background-color: #f1f3f5;
-    border-right: 1px solid var(--border-color);
-    display: flex;
-    flex-direction: column;
-}
-
-.sidebar-header {
-    padding: 1rem;
-    border-bottom: 1px solid var(--border-color);
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.sidebar-header h2 {
-    font-size: 1.1rem;
-    color: var(--accent-color);
-}
-
-.notes-list {
-    flex: 1;
-    overflow-y: auto;
-    padding: 0.5rem;
-}
-
-.note-item {
-    padding: 0.75rem;
-    border-bottom: 1px solid var(--border-color);
-    cursor: pointer;
-    transition: background-color 0.2s;
-}
-
-.note-item:hover {
-    background-color: rgba(0, 0, 0, 0.05);
-}
-
-.note-item.active {
-    background-color: rgba(74, 111, 165, 0.1);
-    border-left: 3px solid var(--primary-color);
-}
-
-.note-item h3 {
-    font-size: 0.95rem;
-    margin-bottom: 0.25rem;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.note-item p {
-    font-size: 0.8rem;
-    color: #666;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.note-editor {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    background-color: white;
-}
-
-.editor-header {
-    padding: 1rem;
-    border-bottom: 1px solid var(--border-color);
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.note-title-input {
-    font-size: 1.2rem;
-    border: none;
-    outline: none;
-    width: 70%;
-    font-weight: bold;
-    color: var(--dark-color);
-}
-
-.editor-actions {
-    display: flex;
-    gap: 0.5rem;
-}
-
-#note-content {
-    flex: 1;
-    padding: 1rem;
-    border: none;
-    outline: none;
-    resize: none;
-    font-size: 1rem;
-    line-height: 1.6;
-}
-
-.primary-btn {
-    background-color: var(--primary-color);
-    color: white;
-    border: none;
-    padding: 0.5rem 1rem;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 0.9rem;
-    transition: background-color 0.2s;
-}
-
-.primary-btn:hover {
-    background-color: var(--accent-color);
-}
-
-.danger-btn {
-    background-color: var(--danger-color);
-    color: white;
-    border: none;
-    padding: 0.5rem 1rem;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 0.9rem;
-    transition: background-color 0.2s;
-}
-
-.danger-btn:hover {
-    background-color: #bd2130;
-}
-
-.empty-state {
-    text-align: center;
-    padding: 2rem;
-    color: #6c757d;
-    font-style: italic;
-}
-
-footer {
-    background-color: var(--primary-color);
-    color: white;
-    text-align: center;
-    padding: 0.5rem;
-    font-size: 0.8rem;
-}
