--- examples/simple_notepad/frontend/index.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Simple Notepad</title>
-    <link rel="stylesheet" href="css/style.css">
-</head>
-<body>
-    <div class="container">
-        <header>
-            <h1>Simple Notepad</h1>
-            <p>A PyWebview Example Application</p>
-        </header>
-        
-        <main>
-            <div class="notes-container">
-                <div class="notes-sidebar">
-                    <div class="sidebar-header">
-                        <h2>My Notes</h2>
-                        <button id="new-note-btn" class="primary-btn">New Note</button>
-                    </div>
-                    <div id="notes-list" class="notes-list">
-                        <!-- Notes will be listed here -->
-                        <div class="empty-state">No notes yet. Create one!</div>
-                    </div>
-                </div>
-                
-                <div class="note-editor">
-                    <div class="editor-header">
-                        <input type="text" id="note-title" placeholder="Note Title" class="note-title-input">
-                        <div class="editor-actions">
-                            <button id="save-note-btn" class="primary-btn">Save</button>
-                            <button id="delete-note-btn" class="danger-btn">Delete</button>
-                        </div>
-                    </div>
-                    <textarea id="note-content" placeholder="Write your note here..."></textarea>
-                </div>
-            </div>
-        </main>
-        
-        <footer>
-            <p>Built with PyWebview - Python + Web Technologies</p>
-        </footer>
-    </div>
-
-    <script src="js/app.js"></script>
-</body>
-</html>
